+++ src/webview/constants/Defaults.ts
@@ -1,3 +1,4 @@
+import { TitleBarNames } from "../models";
 import { Gradients } from "./Gradients";
 
 export const Defaults = {
@@ -9,5 +10,5 @@ export const Defaults = {
   height: 675,
   shadow: 25,
   outerBackground: Gradients[3],
-  showTitleBar: false,
+  titleBarType: TitleBarNames.None,
 }