+++ src/webview/models/TitleBarType.ts
@@ -1,0 +1,10 @@
+
+export const TitleBarNames = {
+  None: "None",
+  macOS: "macOS",
+  macOSInactive: "macOS (inactive)",
+  Windows: "Windows",
+  WindowsInactive: "Windows (inactive)",
+}
+
+export type TitleBarType = keyof typeof TitleBarNames;