+++ src/webview/models/index.ts
@@ -1,2 +1,3 @@
 export * from './FontType';
 export * from './ScreenshotDetails';
+export * from './TitleBarType';